+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Fixed bug where commands only worked for last loaded webview
 - Store note state (edit/settings etc) to bring back note state when viewing previously unloaded/hidden webview
 - Added markdown quick reference to settings pane
+- Allow user to resize settings pane
 
 # 0.2.3